libft-unit-tests, ftmemmove, ftisascii, fttoupper

+++ libft/ft_memcpy.c
@@ -6,15 +6,15 @@
 /*   By: rbetz <rbetz@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/03/23 12:09:34 by rbetz             #+#    #+#             */
-/*   Updated: 2022/03/24 19:12:20 by rbetz            ###   ########.fr       */
+/*   Updated: 2022/03/25 12:02:27 by rbetz            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-void	*ft_memcpy(void *dst, const void *src, size_t n)
+void	*ft_memcpy(void *dst, const void *src, unsigned int n)
 {
-	char	*d;
-	char	*s;
-	int		i;
+	char				*d;
+	char				*s;
+	unsigned int		i;
 
 	d = (char *)dst;
 	s = (char *)src;
